Initial commit of the 'wakeonarp' program
Wakeonarp allows you to send a magic packet to wake a machine
on your network if a arp request for a the IP of this machine
is detected on the local network. It uses libpcap to detect
ARP request and use syslog to log emission of magic packet on
the network. The program can be daemonize to run in background.